Backend optimization (#199)
* Update record creation to use threading

* Add database index

* optimize minelist query

+++ python-backend/app/commands.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 import random
+from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import click
 import names
@@ -37,31 +38,63 @@ def register_commands(app):
     # in terminal you can run $flask <cmd> <arg>
     @app.cli.command()
     @click.argument('num')
-    def create_data(num):
-        party = None
-        for _ in range(int(num)):
-            # Ability to add previous party to have multiple permittee
-            prev_party_guid = party.party_guid if party else None
-            mine_identity = MineIdentity.create_mine_identity(DUMMY_USER_KWARGS)
-            MineDetail.create_mine_detail(mine_identity, generate_mine_no(), generate_mine_name(), DUMMY_USER_KWARGS)
-            MineLocation.create_mine_location(mine_identity, random_geo(), DUMMY_USER_KWARGS)
-            party = Party.create_party(names.get_first_name(), names.get_last_name(), DUMMY_USER_KWARGS)
+    @click.argument('threading', default=True)
+    def create_data(num, threading):
+        """
+        Creates dummy data in the database. If threading=True
+        Use Threading and multiprocessing to create records in chunks of 100.
 
-            create_multiple_mine_tenure(random.randint(0, 4), mine_identity)
-            create_multiple_permit_permittees(random.randint(0, 6), mine_identity, party, prev_party_guid)
+        :param num: number of records create
+        :param threading: use threading or not
+        :return: None
+        """
+        if threading:
+            with ThreadPoolExecutor() as executor:
+                num = int(num)
+                number_batches = []
+                if(num >= 100):
+                    number_batches = [100 for _ in range(100, num, 100)]
+                number_batches.append(num % 100) if num % 100 is not 0 else None
 
-        try:
-            db.session.commit()
-            click.echo(f'Created {num} random mines with tenure, permits, and permittee.')
-        except DBAPIError:
-            db.session.rollback()
-            click.echo(f'Error, failed on commit.')
-            raise
+                task_list = []
+                for batch in number_batches:
+                    task_list.append(executor.submit(_create_data, batch))
+                for task in as_completed(task_list):
+                    try:
+                        data = task.result()
+                    except Exception as exc:
+                        print(f'generated an exception: {exc}')
+        else:
+            _create_data(num)
+
+    def _create_data(num):
+        with app.app_context():
+            party = None
+            for _ in range(int(num)):
+                # Ability to add previous party to have multiple permittee
+                prev_party_guid = party.party_guid if party else None
+                mine_identity = MineIdentity.create_mine_identity(DUMMY_USER_KWARGS)
+                MineDetail.create_mine_detail(mine_identity, generate_mine_no(), generate_mine_name(),
+                                              DUMMY_USER_KWARGS)
+                MineLocation.create_mine_location(mine_identity, random_geo(), DUMMY_USER_KWARGS)
+                party = Party.create_party(names.get_first_name(), names.get_last_name(), DUMMY_USER_KWARGS)
+
+                create_multiple_mine_tenure(random.randint(0, 4), mine_identity)
+                create_multiple_permit_permittees(random.randint(0, 6), mine_identity, party, prev_party_guid)
+
+            try:
+                db.session.commit()
+                print(f'Created {num} random mines with tenure, permits, and permittee.')
+            except DBAPIError:
+                db.session.rollback()
+                raise
 
     @app.cli.command()
     def delete_data():
         meta = db.metadata
         for table in reversed(meta.sorted_tables):
+            if 'view' in table.name:
+                continue
             db.session.execute(table.delete())
         # Reseed Mandatory Data
         PermitStatusCode.create_mine_permit_status_code('O', 'Open permit', 10, DUMMY_USER_KWARGS)
